Make the black actually use black ascii BG colors.
The generated QR codes in the original version didn't work when using a
terminal with black text on white background.

This change makes the library use an actual color code for black (or at
least dark grey) background for the black color, just as it was using
the color code for white.